[RUM] move back `addCustomEvent` logic in rum.ts
